requirements.txt
# J.A.R.V.I.S. - Requirements
# Assistant Vocal Intelligent Local

# Core dependencies
requests==2.31.0
openai==1.3.0

# Speech and Audio
speech-recognition==3.10.0
pyttsx3==2.90
pygame==2.5.2
pyaudio==0.2.11

# Computer Vision and OCR
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
pyautogui==0.9.54

# AI and Machine Learning
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
tortoise-tts==2.8.0

# Data processing
numpy==1.24.4
scipy==1.11.4

# System utilities
psutil==5.9.6

# Optional dependencies for enhanced features
# soundfile==0.12.1
# librosa==0.10.1